Added speedup plotting

+++ utilities.cpp
@@ -46,7 +46,7 @@ int pick_candidate(std::vector<double> probabilities){
     return i;
 }
 
-void log(int i, int pop_size, int nw, int time, int n_nodes){
+void log(int i, int pop_size, int nw, long time, int n_nodes){
     std::cout << "#" << i << "; pop_size: " << pop_size << "; nw: " << nw << "; #nodes: " << n_nodes << "; time: " << (double)time/1000000 << "; thr: " << ((double)i/(double)time)*1000000 << "\r";
     std::cout.flush();
 }
